+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 This release adds an environment check in `@sentry/nextjs` for Vercel deployments (using the `VERCEL_ENV` env variable), and only enables `SentryWebpackPlugin` if the environment is `production`. To override this, [setting `disableClientWebpackPlugin` or `disableServerWebpackPlugin` to `false`](https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/#disable-sentrywebpackplugin) now takes precedence over other checks, rather than being a no-op. Note: Overriding this is not recommended! It can increase build time and clog Release Health data in Sentry with inaccurate noise.
 
+- fix(integrations): Don't add empty stack trace in `RewriteFrames` (#5625)
 - fix(nextjs): Don't run webpack plugin on non-prod Vercel deployments (#5603)
 
 ## 7.11.1